backend/requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
numpy==1.24.3
opencv-python==4.8.0.74
torch==2.0.1
torchvision==0.15.2
ultralytics==8.0.207
PyYAML==6.0
Pillow==10.0.0
python-dotenv==1.0.0
Werkzeug==2.3.6
